requirements.txt
altgraph==0.17.4
appdirs==1.4.4
apturl==0.5.2
attrs==21.2.0
bcrypt==3.2.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cpuset==1.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
fasteners==0.14.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
graphsurgeon==0.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jetson.GPIO==2.1.7
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
onnx-graphsurgeon==0.3.12
packaging==24.2
pandas==1.3.5
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
ply==3.11
protobuf==3.12.4
psutil==7.0.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyodbc==5.2.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
sympy==1.9
systemd-python==234
tensorrt==8.6.2
tensorrt-dispatch==8.6.2
tensorrt-lean==8.6.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uff==0.6.9
ufoLib2==0.13.1
unicodedata2==14.0.0
urllib3==1.26.5
urwid==2.1.2
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
